api/requirements.txt
Flask==3.0.2
flask-sqlalchemy==3.1.1
SQLAlchemy==2.0.28
psycopg2==2.9.9
PyMuPDF==1.23.26
openai==1.14.2
python-dotenv==1.0.1
tensorflow==2.16.1
transformers==4.41.0
tf-keras==2.16.0